(ServerRGBD) use static_cast

+++ rgbd/src/Server_RGBD.cpp
@@ -115,7 +115,7 @@ bool ServerRGBD::serviceCallback(rgbd_msgs::GetRGBDRequest& req, rgbd_msgs::GetR
 
 void ServerRGBD::serviceThreadFunc(const float freq)
 {
-    ros::Rate r(freq);
+    ros::Rate r(static_cast<double>(freq));
     while(nh_.ok())
     {
         cb_queue_.callAvailable();
